refactor + code reformat
