+++ core/src/main/java/com/github/steanky/element/core/BasicElementInspector.java
@@ -1,4 +1,4 @@
-package com.github.steanky.element.core.element;
+package com.github.steanky.element.core;
 
 import com.github.steanky.element.core.factory.FactoryResolver;
 import com.github.steanky.element.core.processor.ProcessorResolver;
