+++ core/src/main/java/com/github/steanky/element/core/ElementFactory.java
@@ -1,5 +1,6 @@
-package com.github.steanky.element.core.element;
+package com.github.steanky.element.core;
 
+import com.github.steanky.element.core.context.ElementContext;
 import com.github.steanky.element.core.dependency.DependencyProvider;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
