+++ core/src/test/java/com/github/steanky/element/core/element/BasicElementInspectorIntegrationTest.java
@@ -1,6 +1,6 @@
 package com.github.steanky.element.core.element;
 
-import com.github.steanky.element.core.ElementException;
+import com.github.steanky.element.core.*;
 import com.github.steanky.element.core.annotation.*;
 import com.github.steanky.element.core.data.BasicDataIdentifier;
 import com.github.steanky.element.core.data.BasicDataInspector;
